Take into account the fact that rdbms are case-sensitive or not and those which are not, the case (either upper or lower) they used to define table/column name
git-svn-id: http://svn.bluexml.com/private/S-IDE@35052 c4e9bb49-730a-0410-b52d-bc9a0859fdfe

+++ FrameworksModules/trunk/com.bluexml.side.Framework.alfresco.sql346E/src/main/java/com/bluexml/side/Integration/alfresco/sql/synchronization/SQLSynchronizationPolicy.java
@@ -83,9 +83,8 @@ public class SQLSynchronizationPolicy implements
 
 	public void onCreateNode(ChildAssociationRef childAssociationRef) {
 		NodeRef nodeRef = childAssociationRef.getChildRef();
-
-		// Only process bluexml nodes
-		if (filterer.accept(nodeRef)) {
+		// Only process declared mapped nodes
+		if (synchroNodeService.exists(nodeRef) && filterer.accept(nodeRef)) {
 			logger.debug("Synchronization policy, CREATE NODE");
 			/*
 			 * Here we cannot determine whether this creation originated from a creation of a new node
@@ -98,14 +97,14 @@ public class SQLSynchronizationPolicy implements
 	}
 
 	public void beforeDeleteNode(NodeRef nodeRef) {
-		if (filterer.accept(nodeRef)) {
+		if (synchroNodeService.exists(nodeRef) && filterer.accept(nodeRef)) {
 			logger.debug("Synchronization policy, DELETE NODE");
 			synchroNodeService.delete(nodeRef);
 		}
 	}
 
 	public void onUpdateProperties(NodeRef nodeRef, Map<QName, Serializable> before_properties, Map<QName, Serializable> after_properties) {
-		if (filterer.accept(nodeRef)) {
+		if (synchroNodeService.exists(nodeRef) && filterer.accept(nodeRef)) {
 			Map<QName, Serializable> changes = new HashMap<QName, Serializable>();
 			for (QName key : after_properties.keySet()) {
 				if (filterer.acceptPropertyQName(key)) {
